Big photo refactorung.
