+++ js/get-thumbnail.js
@@ -1,27 +1,14 @@
-import { openPhoto } from './big-photo.js';
-
-
 const thumbnailTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
 
 /**
- * Обрабатывает событие нажатия на превью фотографии.
- * @param {object} evt
- */
-const onThumbnailClick = (evt, photoData) => {
-  evt.preventDefault();
-  openPhoto(photoData);
-};
-
-
-/**
  * Создаёт разметку превью фотографии.
- * @param {object} photoData Фотография.
+ * @param {object} photoData Данные фотографии.
  * @returns {object} Превью фотографии.
  */
 const getThumbnail = (photoData) => {
   const thumbnail = thumbnailTemplate.cloneNode(true);
-  thumbnail.dataset.id = photoData.id;
+  thumbnail.dataset.photoId = photoData.id;
 
   const thumbnailPicture = thumbnail.querySelector('.picture__img');
   thumbnailPicture.src = photoData.url;
@@ -32,10 +19,6 @@ const getThumbnail = (photoData) => {
 
   const thumbnailLikesCount = thumbnail.querySelector('.picture__likes');
   thumbnailLikesCount.textContent = photoData.likes;
-
-  thumbnail.addEventListener('click', (evt) => {
-    onThumbnailClick(evt, photoData);
-  });
 
   return thumbnail;
 };
